+++ previous-versions/accounting/src/codataccounting/models/operations/create_direct_cost.py
@@ -13,9 +13,12 @@ from typing import Optional
 @dataclasses.dataclass
 class CreateDirectCostRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     connection_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'connectionId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a connection."""
     direct_cost: Optional[shared_directcost.DirectCost] = dataclasses.field(default=None, metadata={'request': { 'media_type': 'application/json' }})
     timeout_in_minutes: Optional[int] = dataclasses.field(default=None, metadata={'query_param': { 'field_name': 'timeoutInMinutes', 'style': 'form', 'explode': True }})
+    r"""Time limit for the push operation to complete before it is timed out."""
     
 
 
@@ -24,11 +27,14 @@ class CreateDirectCostRequest:
 @dataclasses.dataclass
 class CreateDirectCostResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     create_direct_cost_response: Optional[shared_createdirectcostresponse.CreateDirectCostResponse] = dataclasses.field(default=None)
     r"""Success"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""The request made is not valid."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""