+++ previous-versions/accounting/src/codataccounting/models/operations/get_sales_order.py
@@ -12,7 +12,9 @@ from typing import Optional
 @dataclasses.dataclass
 class GetSalesOrderRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     sales_order_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'salesOrderId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a sales order."""
     
 
 
@@ -21,10 +23,13 @@ class GetSalesOrderRequest:
 @dataclasses.dataclass
 class GetSalesOrderResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""
     sales_order: Optional[shared_salesorder.SalesOrder] = dataclasses.field(default=None)
     r"""Success"""