+++ previous-versions/accounting/src/codataccounting/models/operations/get_tax_rate.py
@@ -12,7 +12,9 @@ from typing import Optional
 @dataclasses.dataclass
 class GetTaxRateRequest:
     company_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'companyId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a company."""
     tax_rate_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'taxRateId', 'style': 'simple', 'explode': False }})
+    r"""Unique identifier for a tax rate."""
     
 
 
@@ -21,10 +23,13 @@ class GetTaxRateRequest:
 @dataclasses.dataclass
 class GetTaxRateResponse:
     content_type: str = dataclasses.field()
+    r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
+    r"""HTTP response status code for this operation"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
     r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    r"""Raw HTTP response; suitable for custom response parsing"""
     tax_rate: Optional[shared_taxrate.TaxRate] = dataclasses.field(default=None)
     r"""Success"""