+++ previous-versions/accounting/src/codataccounting/models/shared/withholdingtaxitems.py
@@ -12,6 +12,8 @@ from decimal import Decimal
 @dataclasses.dataclass
 class WithholdingTaxitems:
     amount: Decimal = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('amount'), 'encoder': utils.decimalencoder(False, False), 'decoder': utils.decimaldecoder }})
+    r"""Amount of tax withheld."""
     name: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('name') }})
+    r"""Name assigned to withheld tax."""