+++ previous-versions/accounting/src/codataccounting/sdkconfiguration.py
@@ -20,8 +20,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '3.0.0'
-    sdk_version: str = '0.30.0'
-    gen_version: str = '2.109.3'
+    sdk_version: str = '0.31.0'
+    gen_version: str = '2.147.0'
+    user_agent: str = 'speakeasy-sdk/python 0.31.0 2.147.0 3.0.0 codat-accounting'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: